working through the emotion wheal

+++ Assets/Scripts/STATICS/NUMBERS.cs
@@ -5,7 +5,7 @@ public static class NUMBERS {
 
 	public static int arrayBlockSize = 8;
 	
-	public static int defaultEmotionFillTime = 1.0f;
-	public static int defaultEmotionTransferTime = 1.0f;
+	public static float defaultEmotionFillTime = 1.0f;
+	public static float defaultEmotionTransferTime = 1.0f;
 
 }
